+++ doc/whats-new.md
@@ -136,6 +136,7 @@ Furthermore, you can now bind any keyboard keys (not just 0-9/A-Z).  See [HotKey
  - speaking either "don't care" or "nothing" will mark off an overworld tile using [Speech Recognition](use.md#speech-recognition)
  - when you get a reminder to consider the magical sword, tracker now briefly highlights its location on the map
  - when you get the 8th triforce piece, reminder that dungeon 9 is unlocked now briefly highlights the location of dungeon 9
+ - there's now reminders for Door Repair Count on the Options Menu
  - when you reset the timer, you immediately get put into the 'place the start spot location icon' popup to mark your start screen
  - you can now middle-click-drag across multiple dungeon rooms to mark them as 'off the map' spots
  - Show/Run Custom button now can 'RUN' URLs, which get launched in your default browser